+++ tests/unit/plugins/relay/api/remote/test-relay-remote-network.cpp
@@ -56,11 +56,11 @@ TEST_GROUP(RelayRemoteNetworkWithRemote)
         /* create two relay remotes */
         ptr_relay_remote = relay_remote_new ("remote",
                                              "http://localhost:9000",
-                                             "off", NULL, "on",
+                                             NULL, "off", NULL, "on",
                                              "secret", "secretbase32");
         ptr_relay_remote2 = relay_remote_new ("remote2",
                                               "https://localhost:9001/",
-                                              "off", "my_proxy", "off",
+                                              "30", "off", "my_proxy", "off",
                                               "secret", "secretbase32");
     }